Pin upstream, correct the spread rule, take G1 down to one blocker
The owner supplied the Bot Kit, the Event Contracts docs and the starter
template. All four sources are now pinned in skills-lock.json, including
@somnia-chain/markets-sdk@0.29.0 — the real SDK. Every earlier search failed
because the name was guessed as @dreamdex/*. Guessing a package name is the same
error as guessing an ABI, and it produced the same result: a confident, wrong
answer that read as evidence of absence. D-014 supersedes D-010.

/* §5.2 spread is absolute, not a ratio of mid — D-011 supersedes D-005 */
frontend.md gives two worked examples and D-005 contradicts both:
  §3.2  0.4920/0.5080  spread 0.0160  document says 160 bps  ratio gives  320
  §3.7  0.4700/0.5350  spread 0.0650  document says 650 bps  ratio gives 1293
Pinned IEventContracts.sol corroborates the scale: price is probability in 1e6
units, oneCollateral is one whole contract. maxSpread is now an absolute bound
in the book's own units and widened uint32 -> uint128. Rendering bps for a
reader needs oneCollateral, which is not a field of a stored sample, so that
conversion lives outside the evaluator as display-only and is covered by a test
that reproduces both numbers above.

/* §26 K1 and K3 both stand down — D-012 */
The book emits OrderPlaced/OrderFilled/OrderRested/OrderCancelled, so Path R has
an event to subscribe to. OrderFilled carries only order ids, but OrderPlaced
carries placedOrder.owner, so per-address attribution is possible by correlating
the two. Neither kill criterion fires.

/* §22 G1 */
probe:dreamdex now exits zero against live Shannon testnet: chain 50312, venue
addresses read from the SDK at runtime, 38 MarketCreated logs discovered, 6 live
markets, configured market confirmed. Somnia caps eth_getLogs at 1000 blocks, so
discovery walks backwards in windows and counts failed ones rather than
swallowing them — "no markets found" and "we could not look" are different
answers. G1 is 7 of 8 checks; the last needs the Somnia reactivity reference,
the one item in §0.2's pinning list not supplied.

/* the stale-fixture trap, closed */
It bit twice: once after the ABSENT precedence fix, once after this spread
change. Both times a bare forge test reported a divergence that was only an old
expectation replayed against new code. test:differential now deletes its vectors
in a finally, so they exist only inside the run that made them.

/* not fixed, because it is not ours to fix — D-013, OWNER DECISION */
frontend.md §3.7 specifies a copyable command using dream-rpc.shannon.somnia
.network, which does not resolve. The working endpoint is dream-rpc.somnia
.network. G7 has a stranger paste that command, so it fails a gate rather than
merely reading oddly.

+++ contracts/src/AssizeRegistry.sol
@@ -41,7 +41,7 @@ contract AssizeRegistry {
     struct Commitment {
         address maker;
         bytes32 marketId;
-        uint32 maxSpread;
+        uint128 maxSpread;
         uint128 minSize;
         uint64 start;
         uint64 end;
@@ -100,7 +100,7 @@ contract AssizeRegistry {
         uint256 indexed commitmentId,
         address indexed maker,
         bytes32 indexed marketId,
-        uint32 maxSpread,
+        uint128 maxSpread,
         uint128 minSize,
         uint64 start,
         uint64 end,
@@ -199,7 +199,7 @@ contract AssizeRegistry {
     /// leave the bond exposed to nothing at all while still calling itself a bond.
     function publishCommitment(
         bytes32 marketId,
-        uint32 maxSpread,
+        uint128 maxSpread,
         uint128 minSize,
         uint64 start,
         uint64 end
